refactor: relocate service action schemas
