+++ ciris_engine/sinks/base_sink.py
@@ -7,7 +7,7 @@ import logging
 from typing import Any, Dict, Optional, List
 from dataclasses import asdict
 from abc import ABC, abstractmethod
-from .action_types import ActionType, ActionMessage
+from ciris_engine.schemas.service_actions_v1 import ActionType, ActionMessage
 from ..registries.circuit_breaker import CircuitBreakerError
 
 logger = logging.getLogger(__name__)
